refactor(panel): extract simulation orchestration to state/simulation.ts
Pulls simulation, simNodes, replaceSimulation, wakePhysics, tick,
syncRenderedPositionsFromSimulation, and the settled-physics gate
(SETTLED_THRESHOLD / SETTLE_EPSILON_SQ / WAKE_ALPHA) out of index.ts
and into a single state module that owns the simulation lifecycle.

Mechanical extraction. Module returns a SimulationState handle with:
  tick()                              — gated; advances + lerps + flushes
  wakePhysics()                       — reset gate + bump alpha
  replaceActive(opts)                 — was replaceSimulation()
  syncRenderedPositionsFromSimulation()
  primeOnce()                         — single sim tick before display
  getNodePosition(idx)                — replaces direct simNodes[idx].x reads
  getSimNodes()                       — for snapshot save
  dispose()                           — was simulation.stop()

renderedPositions becomes private to the module (was an index.ts let).
The two external readers (sidePanel onNavigate, search onResult) now go
through getNodePosition; updateOnboardingCamera reads nodePositions
instead, which tracks the same per-tick smoothed values.

setupGraph re-creates the SimulationState alongside NodeLayer on graph
reload — it captures graph + nodes + edges by reference, so its
lifetime is bound to the graph instance.

This is the prerequisite for moving physics to a Web Worker. With the
boundary clean, the worker swap becomes a localized change inside
state/simulation.ts (proxy + worker module behind the same handle
shape) instead of touching ~10 sites in index.ts.
